Handle --lmms-dir
This option allows the user to specify the lmms directory to retrieve
some samples and soundfont file he/she puts.
Indeed, the files coming from LMMS directories and sub-directories
are given by relative path (relative to the LMMS directory) instead of asbsolute path.

With this option, the program can try to retrieve some files placed in the specified directory.