Decompress .mmpz file
- Integrate the use of .mmpz file as a project file input.
- Set a custom LMMS executable (option `--lmms-exe`).

The `--lmms-exe` allows the user to set a custom path of an LMMS executable
that will be used for decopression.
This options is not necessary but can be very useful if you cannot just execute `lmms`
(in general, it could be executed if the program is set in the $PATH variable).