Implement compression of the package on Windows with Zip Utils
The code must be portable on Windows and Linux, so the external dependencies
must be easily used in those platforms. However, I wasted my time trying to compile
libzippp (this library is multi-platform) on Windows.

Maybe I was doing it wrong bu honestly I don't care, so on Linux I uses libzippp
(it is so easy to get things done on alinux distribtution) and on Windows I use
Zip Utils. I have two zip libraries to maintain inside the packager but I can handle it.

Fuck Windows!

Link:
- http://www.wischik.com/lu/programmer/zip_utils.html