Update the project file in order to handle GHC and Zip Utils
GHC is an iplementation of the C++17 filesystem usable in C++14.

Zip Utils is a simple library with a clean API to (de)compress files.
It was designed for Windows systems but can also work on Linux distributions
(I could compile a program using this library). However, I ckeep this use of the library
to Windows because this the library I could compile on it without wastingmy f****** time
trying to compile it properly (even if I prefer libzippp).

The libzippp files are configured to not be compiled by CodeBlocks.

Links:

- https://github.com/gulrak/filesystem
- http://www.wischik.com/lu/programmer/zip_utils.html