Fix package checking on Windows
For some weird reasons, the application cannot check the project file (.mmp)
in the compressed package properly. This is because the underlying libray that
analyses the zip has no information about the uncompressed size of the project file.
This happens only on Windows.

This commit fixes that by checking the uncompressed size of the .mmp
and set the size of the buffer used to store the content of the uncompressed project
to a value big enough to handle big projects (4 Mio).

NB: The biggest project I saw takes almost 2 Mio on the hard drive.