Improve the LMMS project file checking
In the previous version the app verified if the project file is valid as an XML document.
This is a problem because just because an XML document is valid does not means it is the exact document
the user can expect.

Indeed, LMMS can handle several types of data depending on what type of document is specified in the XML file.
In LMMS, you can have an XML that describes an instrument preset/setting, or the entire project.
Because of that, a package that just contains a instrument setting (an .xpf file) can be considered as a valid package.

This commit fixes that issue by checking directly the content of the project document.

Here is a example of a valid project:

```
<?xml version="1.0"?>
<!DOCTYPE lmms-project>
<lmms-project creator="LMMS" version="1.0" creatorversion="1.2.2" type="song">
    <!-- Some data -->
    ...
    <!-- Some data -->
</lmms-project>
```

A valid project is a song project generated by a valid version of LMMS.

Notes:

- Only stable versions of LMMS are supported, due to stability concerns.
- For now, only LMMS 1.2 versions are supported.
- The list of supported LMMS versions will be updated when a new stable release of the DAW is published.