Implement --verbose option
It is now possible to use the `--verbose` option in order to get exhaustive messages in the terminal.
Only important messages (relevant information and error messages) are displayed.

Note:

`--verbose` can only be used after you specified the operation flag and the path.
Using this option before any operation type flag is not valid.
This is a knwon bug that will be fixed when a proper argument parser will be integrated.