Fix missing resource file to export
LMMS sets resources in its project files by using their absolute path.
However, it is possible to manually set the files by using their relative path.

If the path of those files is relative to the directory of the project, the copy cannot be processed.

Example:

a/
|- project.mpp
|- audio.wav

The WAVE file is specified by ./audio.wav, or simply audio.wav instead of /home/toto/data/a/audio.wav
In this case, the file cannot be exported to the package.

This commit fixes that by putting the project directory into the resource directories.